Added test descriptions for rewards; extracted helper functions from tests
